Connect Refresh: Login – Move and update terms-of-service as a sub-heading
